V3.0.11 - Added IC2 and Addons
Previous commit version was incorrect, was supposed to be V3.0.10.

Please note that Compact Windmills has been disabled due to a missing
mod requirement. Will be enabled and tested once missing mod is
installed.

Also, Compact Watermills mod currently does not appear to be named
correctly in-game. Use NEI's drop-down list to search for the mod, or
type "tile." in the search bar. The watermills should be the last set of
machines as of this version.

Mods added:
- IndustrialCraft 2 Experimental: 2.1.478
- Advanced Machines: (no version number)
- Compact Solars: 4.4.28.310
- CompactWindmills: 1.0.3.1 (disabled due to missing mod requirement)
- Compact Watermills: 0.8.0

Config changes:
- ArmorStatusHUD: Max damage display on items now hidden