V3.0.1.12 - Stage 11 Mod Installation
Newly installed mods:
- Another One Bites the Dust: 2.1.0
- Mantle: 0.3.0
- Tinkers' Construct: 1.6.0d33
- TiC Tooltips: 1.1.8a

Notes:
- Previous commit version number incorrect, was supposed to be V3.0.1.11
- Another One Bites the Dust (AOBD) has an updated version, however it
currently crashes the client. Reverting to older version for now
- Changed TinkersConstruct config to automatically provide tier 1
patterns when crafting and placing a pattern chest.
- Changed TinkersConstruct config so that players now drop heads upon
death.
- "Mod List.xlsx" file now updated to contain all currently installed
mods, as well as possible mods which will be considered for installation
into the modpack. Please pay attention to this file if you wish to know
what is currently installed in the modpack (installed mods have an "X"
next to their name).