V3.1.0.0 - O-Week Update
Update for O-Week, as well as using a new server!  Updated mods are too
numerous to list, however the following config changes have been
implemented:
- Demon town summoning from Blood Magic has been set to creative only.
If you wish to use this in singleplayer, either switch to creative, or
disable the creative only option in the configs.
- Natura clouds disabled in everywhere except the Nether. Was quite
annoying to see clouds in the end.
- Attempted to disable spawner controller from QuarryPlus, as it is too
OP for the modpack. Has not been tested.
- Disabled Project Red volcanoes, as it can be absolutely horrid to look
at. If you wish to obtain basalt from these volcanoes, an NPC is planned
to sell it to you within the server spawn town.